requirements.txt
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
cfgv==3.4.0
click==8.1.8
decorator==5.2.1
distlib==0.3.9
Django==5.2
django-autoslug==1.9.9
django-cities-light==3.10.1
django-cors-headers==4.7.0
django-debug-toolbar==5.1.0
django-extensions==4.0
django-phonenumber-field==8.0.0
django-rq==3.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
executing==2.2.0
filelock==3.18.0
identify==2.6.9
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
nodeenv==1.9.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
pre_commit==4.2.0
progressbar2==4.5.0
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.9.0
python-dotenv==1.1.0
python-utils==3.9.1
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
rq==2.3.1
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.13.1
Unidecode==1.3.8
virtualenv==20.30.0
wcwidth==0.2.13
